==> notebooks/00_reference_setup.ipynb <==
### Description

This file sets up the dictionary of names and random seeds for use throughout the analysis

### Inputs

# =========================================
# For accessing directories
# =========================================
root_dir = "/local/data/artemis/workspace/jfs2167/recon_eval" # Set this to the path of the project
reference_output_dir = f"{root_dir}/references"

### Modules

from collections import defaultdict
import numpy as np
import pickle

### Creating reference dictionaries and arrays

# =========================================
# Ensemble lists
# =========================================
mems_dict = {} 
mems_dict['CanESM2'] = ['r1r10','r1r9','r3r1','r3r8','r4r5','r5r10','r1r4','r2r1','r3r2','r3r9','r4r6','r5r2','r1r6','r2r2','r3r4','r4r1','r4r7','r5r4','r1r7','r2r3','r3r6','r4r10','r4r8','r5r5','r1r8','r2r8','r3r7','r4r3','r5r1','r5r9']
mems_dict['CESM'] = ['001','002','009','010','011','012','013','014','015','016','017','018','020','021','023','024','025','030','031','034','035','101','102','103','104']
mems_dict['GFDL'] = ['01','02','03','04','05','06','08','09','10','11','12','13','14','15','16','17','18','19','20','22','23','24','26','27','28','29','30']
mems_dict['MPI'] = ['006','009','014','018','020','021','022','024','025','027','028','038','039','041','043','045','046','057','060','064','070','075','077','078','080','081','083','091','095','098']

# =========================================
# Reducing to the 100 member LET
# =========================================
mems_dict_LET = {}
members=[1,2,9,10,11,12,13,14,15,16,17,18,20,21,23,24,25,30,31,34,35,101,102,103,104]
mems_dict_LET["CESM"] = [x for x in mems_dict["CESM"] if int(x) in members]
members=[1,2,3,4,5,6,8,9,10,11,12,13,14,16,17,18,19,20,22,23,26,27,28,29,30]
mems_dict_LET["GFDL"] = [x for x in mems_dict["GFDL"] if int(x) in members]
members=[6,9,14,20,22,24,25,27,38,43,45,46,57,60,64,70,75,77,78,80,81,83,91,95,98]
mems_dict_LET["MPI"] = [x for x in mems_dict["MPI"] if int(x) in members]
members=['r1r10', 'r1r6', 'r1r7', 'r1r9', 'r2r1', 'r2r2', 'r2r8', 'r3r1', 'r3r2', 'r3r4',
        'r3r6', 'r3r7', 'r3r9', 'r4r1', 'r4r3', 'r4r5', 'r4r6', 'r4r7', 'r4r8', 'r5r1', 
        'r5r10', 'r5r2', 'r5r4', 'r5r5', 'r5r9']        
mems_dict_LET["CanESM2"] = [x for x in mems_dict["CanESM2"] if x in members]

# =========================================
# For seeding random states
# =========================================

# Generate a random array of integers of size # total ensemble members
# Use these integers as seeds for the various random parts of the modeling process
num_mems = sum([len(value) for value in mems_dict.values()])
np.random.RandomState(10)
random_seeds = np.random.randint(1000, size=(50,num_mems))

# Creating a dictionary to access the seed column based on ensemble/member names
seed_count = 0
seed_loc_dict = defaultdict(dict)
for ens,ens_list in mems_dict.items():
    for member in ens_list:
        seed_loc_dict[ens][member] = seed_count
        seed_count +=1

### Saving references to a directory

path_mems = f"{reference_output_dir}/members_dict.pickle"
path_LET = f"{reference_output_dir}/members_LET_dict.pickle"
path_seeds = f"{reference_output_dir}/random_seeds.npy"
path_loc = f"{reference_output_dir}/members_seed_loc_dict.pickle"

with open(path_mems, 'wb') as handle:
    pickle.dump(mems_dict, handle)
    
with open(path_LET, 'wb') as handle:
    pickle.dump(mems_dict_LET, handle)
    
np.save(path_seeds, random_seeds)

with open(path_loc, 'wb') as handle:
    pickle.dump(seed_loc_dict, handle, protocol=pickle.HIGHEST_PROTOCOL)

# Code for loading in these files
reference_output_dir = "/local/data/artemis/workspace/jfs2167/recon_eval/references"

path_mems = f"{reference_output_dir}/members_dict.pickle"
path_LET = f"{reference_output_dir}/members_LET_dict.pickle"
path_seeds = f"{reference_output_dir}/random_seeds.npy"
path_loc = f"{reference_output_dir}/members_seed_loc_dict.pickle"

with open(path_mems,'rb') as handle:
    mems_dict = pickle.load(handle)

with open(path_LET,'rb') as handle:
    LET_dict = pickle.load(handle)
    
random_seeds = np.load(path_seeds)    
    
with open(path_loc,'rb') as handle:
    seed_loc_dict = pickle.load(handle)

path_mems = f"{reference_output_dir}/members_dict.pickle"
path_LET = f"{reference_output_dir}/members_LET_dict.pickle"
path_seeds = f"{reference_output_dir}/random_seeds.npy"
path_loc = f"{reference_output_dir}/members_seed_loc_dict.pickle"

with open(path_LET, 'wb') as handle:
    pickle.dump(mems_dict_LET, handle)
